ActiveRecord module for using MuchSlug with ActiveRecord records
This logic is being extracted from Ardb.  Ardb is designed to only
work with ActiveRecord but MuchSlug is more generic and is
theoretically usable with any ORM.  Thus I've broken out all of
the ActiveRecord-specific logic into a MuchSlug::ActiveRecord
mixin and I've commonized the logic that is ORM-agnostic into
singleton methods on the MuchSlug module.  This sets us up to
extend this API for use with Sequel, etc records.

Otherwise all of the logic should be fairly identical to the Ardb
version.  Here are some notes on the things that are different:

* I broke out some of the common logic into their own files since
  we have a fresh namespace here to work with (ie. MuchSlug::Slug).
* I moved the tests `NON_WORD_CHARS` into the Factory since it is
  used across multiple tests files.
* I moved the `DEFAULT_*` constants to singleton methods since they
  are used in multiple files now.
* I don't unit test the non `default_*` methonds in `MuchSlug`, I
  count on the ActiveRecord module tests to test them for me.  So
  the ActiveRecord tests are acting kinda like system tests, FYI.

Note: I plan on using this in the Rails testing app we are building
out.  I'll test this in a more production/live scenario when I
start using this in that app.

Note also: I'm planning on removing `Ardb::HasSlug` and all related
apis from Ardb.  This will be a breaking change for Ardb but I'm
less concerned about the ramifications of this b/c of our new
situation.