switch the default preprocessor to `:to_s`
This is a better default as it does nothing by default. We don't
want to assume that `:downcase` is a good slug convention for all
users.

In the future it would be nice to add some global configuration to
allow setting a new default preprocessor if desired. The same is
true for the default attribute name and default separator.