rework MuchSlug::Configs as MuchSlug::HasSlugRegistry
This uses a formal object instead of just relying on the Hash
primitive. This allows for specific tests against specific
attributes but also allows moving the registry logic out of the
`set_has_slug` singleton method and into the Registry object.
This better encapsulates and names the logic.

In addition, this did a number of cleanups to make the gem ready
for general consumption:

* switch to modern keyword argument usage
* default the `allow_underscores` setting to `true` since this is
  a more appropriate default given the principle of least surprise
* use strings for the registry attribute name values b/c Strings
  are more universal