rework updating slug values; remove `reset_slug` API
This reworks how MuchSlug updates slug values to keep them in
sync with the source value. Previously, MuchSlug would only update
the slug value if the slug attribute was blank. MuchSlug provided a
`reset_slug` API to make the slug attribute blank and then relied
on its late-bound `after_update` callback to re-sync the slug
value.

This has a number of confusing edge cases:

* you have to remember to use `reset_slug`
* you have to call reset slug in a `before_update` callback or
  something that runs after validation but before `after_update`
* you can't call `reset_slug` before validations b/c the blank
  slug will fail validation

The overall goal is that MuchSlug syncs slug values with their
source in a late-bound way. The sync needs to happen after the
save transaction so that users can use db-generated values (like
IDs) in the slug source.

This switches to always calculating the new slug value and then
only updating the record if the generated value doesn't match the
current value. This removes the need for `reset_slug` and removes
a bunch of edge-case concerns. Now to re-sync a slug, the user
needs only to save the record instance and the late-bound
callbacks will update the slug value only if necessary.

Note: one behavior change with this is that you can no longer
set ad-hoc slug values. The callbacks will always overwrite any
ad-hoc value with the sync'd value from the source. This was a
little-known feature that was never really used anyway.