switch back to not allowing underscores by default
When writing up a new README with usage example I couldn't justify
this. The goal is to consistently make slug values be URL friendly
and having it replace underscores with the separator feels better
since underscores act as word-barriers.

I think a better long-term solution is to let users configure the
global defaults themselves and not be stuck with whatever base
defaults MuchSlug happens to choose.