Update to work with STI records
This updates the `MuchSlug::ActiveRecord` to have an `inherited`
hook that copies registry entries to any classes that inherit
from it. This makes it so a `has_slug` configuration can be set
in a base class but still take effect in a child class. Without
this, any child classes will not run the logic to generate their
slug and will leave it blank.